Completed file restructuring to separate three different sections of expense tracker to their own respective folders and components.
