+++ frontend/components/SavingsComponent/SavingsForm.jsx
@@ -1,0 +1,99 @@
+import {useState, React} from "react";
+import RenderSavings from "./RenderSavings";
+
+export default function SavingsForm() {
+  // New Saving Goal
+  const [goals, setGoals] = useState([]);
+  const [description, setDescription] = useState("");
+  const [amount, setAmount] = useState("");
+  const [date, setDate] = useState(new Date());
+
+  const handleCreateGoal = (e) => {
+    e.preventDefault();
+    const newGoal = {
+      description: description,
+      amount: amount,
+      date: date,
+    };
+    console.log(newGoal);
+
+    setGoals([...goals, newGoal]);
+    setDescription("");
+    setAmount("");
+    setDate(getDate(new Date()));
+  };
+
+  return (
+    <>
+      <div className="">
+        <div className="w-full max-w-sm p-4 bg-white border border-gray-200 rounded-lg shadow sm:p-6 md:p-8 dark:bg-gray-800 dark:border-gray-700">
+          <form className="space-y-6" action="#" onSubmit={handleCreateGoal}>
+            <h5 className="text-xl font-medium text-gray-900 dark:text-white">
+              New Saving Goal
+            </h5>
+
+            <div className="mb-5">
+              <label
+                htmlFor="base-input"
+                className="block mb-2 text-sm font-medium text-gray-900 dark:text-white"
+              >
+                Description
+              </label>
+              <input
+                type="text"
+                id="base-input"
+                className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+                value={description}
+                onChange={(e) => setDescription(e.target.value)}
+              />
+            </div>
+
+            <div className="grid gap-6 mb-6 md:grid-cols-2">
+              <div>
+                <label
+                  htmlFor="amount"
+                  className="block mb-2 text-sm font-medium text-gray-900 dark:text-white"
+                >
+                  Amount
+                </label>
+                <input
+                  type="text"
+                  id="amount"
+                  className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+                  placeholder="$"
+                  value={amount}
+                  onChange={(e) => setAmount(e.target.value)}
+                  required
+                />
+              </div>
+              <div>
+                <label className="block mb-2 text-sm font-medium text-gray-900 dark:text-white">
+                  Date
+                </label>
+
+                <input
+                  aria-label="Date"
+                  type="date"
+                  className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+                  value={date}
+                  onChange={(e) => setDate(e.target.value)}
+                />
+              </div>
+            </div>
+
+            <button
+              type="submit"
+              className="w-full my-4 text-white bg-green-400 hover:bg-green-500 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
+            >
+              Create Goal
+            </button>
+          </form>
+        </div>
+
+        <div className="mt-10">
+          <RenderSavings />
+        </div>
+      </div>
+    </>
+  );
+}
